A word-frequency report, and the merge sort is a loop over levels
examples/io/freq.oro — `sort | uniq -c | sort -rn` in Oroboros. 158 lines of
code, the largest program in this language by 1.4x, byte-identical to those
tools on seven real files, and 951 of 951 integer operations bounded with no
-checked.

THE TERMINATION FINDING is the one to keep. The four loops of 37 that are not
proven are the merge sort's — `w = 1, 2, 4, …` and `lo += 2w`. Size-change
handles an ascending counter by orientation, μ = −v, and that works because the
step is a CONSTANT. So karatsuba-2026-08-30's rule for eliminating recursion —
a balanced, data-independent recursion is a loop over levels — produces exactly
the loop shape the termination checker cannot prove, and every future FFT,
binary search and level-walk lands here.

TWO COMPILER BUGS, both pre-existing.

A program that COMPUTES a byte could not get a byte buffer, and neither
existing path decides it alone: the syntactic rule knows a copy from a narrowed
table and was closed under nothing, the interval rule knows the arithmetic and
cannot see the copy, because BufferRange runs on the build lambda alone where
the source is free. A report line is one of each. storedRange now recurses
through arithmetic, plus one rule needing no fact about its operand at all —
|a % d| < |d| whatever a is — which is what makes a digit exact. The
anti-circularity guard moved into the recursion, because arithmetic can bury a
self-read.

And a `let` inside a loop guard emitted dead statements that Go refuses. `e.emit`
is not a query: the BCE pass emitted the bound and then decided whether to use
it. Nine lines reproduce it, and no program had ever put a `let` in a loop
guard.

TWO PROGRAM BUGS. The capacity guard exited before counting repetitions of the
last distinct word, so `a a` reported 1. And the word cap silently TRUNCATED —
`the` 433 against 440 on the README — so it is now derived from the byte cap and
cannot be reached: k words need 2k−1 bytes.

The sort is the differential case `merge-sort` on three targets, verified to
fail against two bugs; windows is skipped for a compiler ceiling, third program
to hit it and the first that is nothing but a sort.

G5's portable-layer debt is closed at 0.99x, and gauntlet-2026-09-07 is
corrected: it was one program, not three.

ADR 0020's measurement is re-taken at matched widths — the fix above narrows its
workspace — and 4.5x becomes 1.84x, which is that result's own §1.1 landing on
its own number. The allocation column is unchanged: 0 against 1.

68 of 69 pre-existing emitted files byte-identical; 29 differential cases green.

+++ gauntlet/go/gen_workspace.go
@@ -2,8 +2,8 @@
 
 package gauntlet
 
-func GenMacFresh(a []byte, b []byte) []int {
-	w := make([]int, 65536)
+func GenMacFresh(a []byte, b []byte) []uint16 {
+	w := make([]uint16, 65536)
 	w2 := w
 	var i int = 0
 	var n1 int = 65536
@@ -13,13 +13,13 @@ func GenMacFresh(a []byte, b []byte) []int {
 		if (i >= 65536) {
 			break
 		}
-		w2[i] = (int(s2[i]) * int(s3[i]))
+		w2[i] = uint16((int(s2[i]) * int(s3[i])))
 		continue
 	}
 	return w2
 }
 
-func GenMacInto(w []int, a []byte, b []byte) []int {
+func GenMacInto(w []uint16, a []byte, b []byte) []uint16 {
 	w2 := w
 	var i int = 0
 	var n1 int = 65536
@@ -29,7 +29,7 @@ func GenMacInto(w []int, a []byte, b []byte) []int {
 		if (i >= 65536) {
 			break
 		}
-		w2[i] = (int(s2[i]) * int(s3[i]))
+		w2[i] = uint16((int(s2[i]) * int(s3[i])))
 		continue
 	}
 	return w2
